+++ fabric/src/main/java/com/yanny/ali/fabric/network/NetworkUtils.java
@@ -11,16 +11,16 @@ public class NetworkUtils {
     public static void registerClient() {
         Client client = new Client();
 
-        PayloadTypeRegistry.clientboundPlay().register(LootDataChunkMessage.TYPE, LootDataChunkMessage.CODEC);
-        PayloadTypeRegistry.clientboundPlay().register(StartMessage.TYPE, StartMessage.CODEC);
-        PayloadTypeRegistry.clientboundPlay().register(DoneMessage.TYPE, DoneMessage.CODEC);
-
         ClientPlayNetworking.registerGlobalReceiver(LootDataChunkMessage.TYPE, client::onLootDataChunk);
         ClientPlayNetworking.registerGlobalReceiver(StartMessage.TYPE, client::onStart);
         ClientPlayNetworking.registerGlobalReceiver(DoneMessage.TYPE, client::onDone);
     }
 
     public static void registerCommon() {
+        PayloadTypeRegistry.clientboundPlay().register(LootDataChunkMessage.TYPE, LootDataChunkMessage.CODEC);
+        PayloadTypeRegistry.clientboundPlay().register(StartMessage.TYPE, StartMessage.CODEC);
+        PayloadTypeRegistry.clientboundPlay().register(DoneMessage.TYPE, DoneMessage.CODEC);
+
         PayloadTypeRegistry.serverboundPlay().register(StartMessage.TYPE, StartMessage.CODEC);
 
         ServerPlayNetworking.registerGlobalReceiver(StartMessage.TYPE, (s, c) -> {});