+++ common/src/test/java/com/yanny/ali/test/GenericTooltipTest.java
@@ -902,7 +902,7 @@ public class GenericTooltipTest {
                         "    -> minecraft:tooltip_display",
                         "      -> Hide Tooltip: false",
                         "    -> minecraft:break_sound",
-                        "      -> Sound: minecraft:entity.item.break",
+                        "      -> Value: minecraft:entity.item.break",
                         "    -> minecraft:use_effects",
                         "      -> Can Sprint: false",
                         "      -> Interact Vibrations: true",